iOS 6 beta xcode tweaks and minor fixes detected by new clang
git-svn-id: http://zxing.googlecode.com/svn/trunk@2315 59b500cc-1b3d-0410-9834-0bbf25fbcc57

+++ objc/src/ZXing/ZXCGImageLuminanceSource.mm
@@ -29,8 +29,9 @@
                                        top:(size_t)top
                                      width:(size_t)width
                                     height:(size_t)height {
-  return zxing::CGImageLuminanceSource::createImageFromBuffer
-    (buffer, left, top, width, height);
+  return zxing::CGImageLuminanceSource::createImageFromBuffer(
+    buffer, (int)left, (int)top, (int)width, (int)height
+    );
 }
 
 - (id)initWithZXImage:(ZXImage*)image 
@@ -38,7 +39,9 @@
                   top:(size_t)top
                 width:(size_t)width
                height:(size_t)height {
-  self = [super initWithNative:new zxing::CGImageLuminanceSource(image.cgimage, left, top, width, height)];
+  self = [super initWithNative:new zxing::CGImageLuminanceSource(
+      image.cgimage, (int)left, (int)top, (int)width, (int)height
+      )];
   return self;
 }
 
@@ -52,7 +55,9 @@
                   top:(size_t)top
                 width:(size_t)width
                height:(size_t)height {
-  self = [super initWithNative:new zxing::CGImageLuminanceSource(image, left, top, width, height)];
+  self = [super initWithNative:new zxing::CGImageLuminanceSource(
+      image, (int)left, (int)top, (int)width, (int)height
+      )];
   return self;
 }
 
@@ -66,7 +71,9 @@
                  top:(size_t)top
                width:(size_t)width
               height:(size_t)height {
-  self = [super initWithNative:new zxing::CGImageLuminanceSource(buffer, left, top, width, height)];
+  self = [super initWithNative:new zxing::CGImageLuminanceSource(
+      buffer, (int)left, (int)top, (int)width, (int)height
+      )];
   return self;
 }
